s3: Optimize catalog iteration performance
S3 supports resuming iteration from a place in the middle of iteration.
This adds support for that. When a user tries to hit the catalog endpoint
this can mean the difference between failure to list and success.

To give an idea of benchmark (where this is the ~300th repo in the list):
$ time curl 'http://localhost:7101/v2/_catalog?n=10&last=brh/ack' -o /dev/null -s

real    0m4.929s
user    0m0.005s
sys     0m0.008s

$ time curl 'http://localhost:7101/v2/_catalog?n=10&last=brh/ack' -o /dev/null -s

real    1m51.244s
user    0m0.005s
sys     0m0.009s

Signed-off-by: Sargun Dhillon <sargun@sargun.me>